+++ src/rinoh/frontend/sphinx/__init__.py
@@ -329,9 +329,9 @@ def default_domain_indices(config):
 
 def deprecation_warnings(config, logger):
     if config.rinoh_stylesheet:
-        logger.warning("'rinoh_stylesheet' has been deprecated."
-                       " Configure the stylesheet in the document template"
-                       " configuration.")
+        logger.warning("Support for 'rinoh_stylesheet' has been removed."
+                       " Instead, please specify the stylesheet to use in your"
+                       " template configuration.")
 
 
 def setup(app):