+++ tests/test_attribute.py
@@ -1,4 +1,13 @@
+# This file is part of rinohtype, the Python document preparation system.
+#
+# Copyright (c) Alex Fargus.
+#
+# Use of this source code is subject to the terms of the GNU Affero General
+# Public License v3. See the LICENSE file or http://www.gnu.org/licenses/.
+
+
 import pytest
+
 from rinoh.attribute import Attribute, Bool
 from rinoh.style import Style
 
@@ -10,20 +19,21 @@ class MockStyle(Style):
 
 def test_attribute_default():
     style = MockStyle()
-    assert(style.boolean == True)
+    assert style.boolean is True
 
 
 def test_attribute_non_default():
     style = MockStyle(boolean=False)
-    assert(style.boolean == False)
+    assert style.boolean is False
 
 
 def test_attribute_from_class():
     attribute = MockStyle.boolean
-    assert(attribute.name == "boolean")
-    assert(attribute.description == "A boolean attribute")
-    assert(attribute.accepted_type == Bool)
-    assert(attribute.default_value == True)
+    assert attribute.name == "boolean"
+    assert attribute.description == "A boolean attribute"
+    assert attribute.accepted_type == Bool
+    assert attribute.default_value is True
+
 
 def test_attribute_wrong_type():
     style = MockStyle()